+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## not released
 
+- Add: Info to restart Joplin on settings change
+
 ## v0.0.2 (2023-12-27)
 
 - Rename plugin